Improved all loads. Added new load formats for a lot of classes. Bug fixes.
Please check what I did before continuing.

+++ CourseUnit.cpp
@@ -6,7 +6,7 @@
 #include "College.h"
 #include "University.h"
 
-CourseUnit::CourseUnit(string n, string a, Course* c, unsigned short int y, unsigned short int s, unsigned int credits)
+CourseUnit::CourseUnit(string n, string a, Course* c, unsigned short int y, unsigned short int s, double credits)
 	: name(n), acronym(a), year(y), semester(s), credits(credits), course(c)
 {
 	c->addCourseUnit(this);
