+++ Student.h
@@ -25,7 +25,7 @@ protected:
 	Tutor* tutor;
 	unsigned short int year;
 	string status; //REGULAR STUDENT, WORKING STUDENT, SPECIAL NEEDS
-	unsigned int credits; //CREDITS THIS YEAR
+	double credits; //CREDITS THIS YEAR
 
 	virtual void assignEmail();
 	virtual void assignID();
@@ -46,7 +46,7 @@ public:
 		
 	//SETS
 	void setStatus(string s) { status = s; }
-	void setCredits(unsigned int c) { credits = c; }
+	void setCredits(double c) { credits = c; }
 	void setYear(unsigned short int y) { year = y; }
 
 	//GETS
@@ -55,7 +55,7 @@ public:
 	map<CourseUnit*, CourseUnitClass*> getClassesCurrentlyAtending() const { return classesCurrentlyAtending; }
 	unsigned short int getYear() const { return year; }
 	string getStatus() const { return status; }
-	unsigned int getCredits() const { return credits; }
+	double getCredits() const { return credits; }
 
 };
